requirements.txt
aiostream==0.5.2
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
Deprecated==1.2.14
dnspython==2.5.0
ecdsa==0.18.0
ecs-logging==2.1.0
elastic-apm==6.21.3
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-pagination==0.12.14
fastapi-utils==0.7.0
firebase-admin==6.4.0
gcloud==0.18.3
google-api-core==2.16.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
instaloader==4.11
jwcrypto==1.5.1
jws==0.1.3
lark==1.2.2
MarkupSafe==3.0.2
msgpack==1.0.7
mypy-extensions==1.0.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
packaging==24.1
pandas==2.2.1
pillow==10.3.0
pluggy==1.5.0
plum-dispatch==1.7.4
priority==2.0.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.6.0
pymongo==4.6.1
pyparsing==3.1.1
pypng==0.20220715.0
PyQRCode==1.2.1
Pyrebase4==4.6.0
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-jwt==3.3.4
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
segno==1.6.1
setuptools==74.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
stripe==11.3.0
taskgroup==0.0.0a4
tomli==2.0.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.0.post1
wrapt==1.14.1
wsproto==1.2.0
zipp==3.18.1
google-cloud-bigquery==3.27.0